card.mk [FEATURE]: environment variables removal
Added CORE_BASE path which simplifies the sourcing of CORE specific
configuration files. Added DMA_TYPE variable with which a DMA type can
be chosen. The default value for this variable is located in the core.mk
file which is also included. USER_CONST renamed to APP_CONF because
there were many USER_CONST files previously and the purpose of this file
was not known anyway.